react adjustments
