+++ package.json
@@ -22,7 +22,9 @@
     "@babel/preset-env": "^7.17.10",
     "@babel/preset-react": "^7.16.7",
     "babel-loader": "^8.2.5",
+    "css-loader": "^6.7.1",
     "html-webpack-plugin": "^5.5.0",
+    "style-loader": "^3.3.1",
     "webpack": "^5.71.0",
     "webpack-cli": "^4.9.2",
     "webpack-dev-server": "^4.7.4"
@@ -35,6 +37,7 @@
     "papaparse": "^5.3.2",
     "react": "^18.1.0",
     "react-chartjs-2": "^4.1.0",
-    "react-dom": "^18.1.0"
+    "react-dom": "^18.1.0",
+    "react-simple-loading": "^0.0.3"
   }
 }
