+++ src/DashboardComponent/Dashboard.css
@@ -1,0 +1,3 @@
+.welcomeContainer {
+    text-align: center;
+}
