+++ src/WelcomeComponent/Welcome.css
@@ -1,0 +1,3 @@
+p.welcomeP {
+    text-align: left;
+}
