+++ src/index.css
@@ -1,0 +1,5 @@
+.indexComponent {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
