+++ src/index.js
@@ -1,26 +1,51 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
+import React from "react";
+import ReactDOM from "react-dom/client";
+import Welcome from './WelcomeComponent/Welcome';
+import Dashboard from './DashboardComponent/Dashboard';
+import './index.css';
 
-import Welcome from './Welcome';
-import Dashboard from './Dashboard';
+const STEPS = {
+  Welcome: "Welcome",
+  Dashboard: "Dashboard",
+}
 
 class App extends React.Component {
   constructor(props) {
     super(props);
-    this.state = { text: null };
+    this.state = { logData: null, step: STEPS.Welcome };
   }
 
-  onRead = (text) => {
-    this.setState({ text: text })
+  onRead = (logData) => {
+    this.setState({ logData, step: STEPS.Dashboard})
   }
 
   render() {
     return (
-      <div style={{ textAlign: 'center' }}>
-        {!this.state.text ? <Welcome onRead={this.onRead} /> : <Dashboard text={this.state.text} />}
+      <div className="indexComponent">
+        {
+          (() => {
+            switch (this.state.step) {
+              case STEPS.Welcome:
+                console.log("** Welcome **")
+                return <Welcome onRead={this.onRead}/>
+              case STEPS.Dashboard:
+                console.log("** Dashboard **")
+                return <Dashboard logData={this.state.logData}/>
+              default:
+                return <div> ... </div>;
+            }
+          })()
+
+        }
       </div>
     )
   }
 }
 
-ReactDOM.render(<App />, document.getElementById('app'))
+const rootElement = document.getElementById("app");
+const root = ReactDOM.createRoot(rootElement);
+
+
+root.render(
+    <App />
+);
